Move implementation of Eventable into dm.rs
It too is about how to get a reference to a DM context.
