+++ src/engine/strat_engine/dm.rs
@@ -4,9 +4,14 @@
 
 // Get ability to instantiate a devicemapper context.
 
+
+use std::os::unix::io::{AsRawFd, RawFd};
 use std::sync::{Once, ONCE_INIT};
 
 use devicemapper::DM;
+
+use super::super::engine::Eventable;
+use super::super::errors::EngineResult;
 
 static INIT: Once = ONCE_INIT;
 static mut DM_CONTEXT: Option<DM> = None;
@@ -18,5 +23,18 @@ pub fn get_dm_context() -> &'static DM {
             Some(ref context) => context,
             _ => panic!("DM_CONTEXT.is_some()"),
         }
+    }
+}
+
+
+impl Eventable for DM {
+    /// Get file we'd like to have monitored for activity
+    fn get_pollable_fd(&self) -> RawFd {
+        self.file().as_raw_fd()
+    }
+
+    fn clear_event(&self) -> EngineResult<()> {
+        self.arm_poll()?;
+        Ok(())
     }
 }
